--- integrations/stripe/src/definitions/channels.ts
@@ -1,7 +1,0 @@
-import { IntegrationDefinitionProps, messages } from '@botpress/sdk'
-
-export const channels = {
-  channel: {
-    messages: { ...messages.defaults },
-  },
-} satisfies IntegrationDefinitionProps['channels']
